+++ src/app/(home)/_components/TextSlider/TextSlider.module.css
