+++ src/app/(home)/_components/TextSlider/index.tsx
